Added 404 responce
I've added a 'not found' error for when a story can't be found.

This fix solves issues with the Exchange autodiscover service. Before, this service was sending a lot of POST requests to /autodiscover.xml, which accidentally caused a 500 error. Now, by using a 'not found' error, we prevent these mistakes, making everything run smoother and more reliably.