Editor bridge appends SB querystring items
Implemented suggestion from @codyworthen to append Storyblok editor querystring items to links on the page. This allows links to be clicked on the page and become editable. This can be configured in the config file.

Fixed replacement DOM content not allowing duplicate IDs. Useful fix when HTML contains SVG data as applications like Illustrator default to <svg id="Layer_1" ...>